Refactoring. Extracting before bed logic to a module.
