+++ lib/babysitter.ex
@@ -1,8 +1,8 @@
 defmodule BabySitter do
 
   alias BabySitter.ScheduleValidation, as: ScheduleValidation
+  alias BabySitter.BeforeBed, as: BeforeBed
 
-  @before_bed_rate 12
   @after_bed_rate 8
   @after_midnight_rate 16
 
@@ -19,28 +19,11 @@ defmodule BabySitter do
 
   defp calculate_by_pay_rate(start_time, hours, bed_time) do
     end_time = start_time + hours
-    before_bed_pay = calculate_before_bed_pay(start_time, end_time, bed_time)
+    before_bed_pay = BeforeBed.calculate_before_bed_pay(start_time, end_time, bed_time)
     after_bed_pay = calculate_bed_time_pay(start_time, end_time, bed_time)
     after_midnight_pay = calculate_midnight_pay(end_time)
 
     before_bed_pay + after_bed_pay + after_midnight_pay
-  end
-
-  defp calculate_before_bed_time_hours(start_time, _, bed_time) when start_time >= bed_time do
-    0
-  end
-
-  defp calculate_before_bed_time_hours(start_time, end_time, bed_time) when end_time <= bed_time do
-    end_time - start_time
-  end
-
-  defp calculate_before_bed_time_hours(start_time, end_time, bed_time) when end_time > bed_time do
-    bed_time - start_time
-  end
-
-  defp calculate_before_bed_pay(start_time, end_time, bed_time) do
-    before_bed_hours = calculate_before_bed_time_hours(start_time, end_time, bed_time)
-    before_bed_hours * @before_bed_rate
   end
 
   defp calculate_bed_time_hours(_, end_time, bed_time) when end_time <= bed_time do
